+++ Sources/Puddles/Puddles.docc/Resources/code-files/03-BasicNavigation/03-BasicNavigation-01-01.swift
@@ -6,13 +6,13 @@ struct RootNavigator: Navigator {
         Root()
     }
 
-    func applyStateConfiguration(_ configuration: StateConfiguration) {
-        // Handle state configurations
+    func applyTargetState(_ state: TargetState) {
+        // Handle target states
     }
 }
 
 extension RootNavigator {
-    enum StateConfiguration: Hashable {
+    enum TargetState: Hashable {
         case reset
         case showPage
     }