+++ Sources/Puddles/Puddles.docc/Resources/code-files/05-Expectations/expectations-01-03.swift
@@ -22,8 +22,8 @@ struct RootNavigator: Navigator {
         }
     }
 
-    func applyStateConfiguration(_ configuration: StateConfiguration) {
-        switch configuration {
+    func applyTargetState(_ state: TargetState) {
+        switch state {
         case .reset:
             path = []
         case .showPage:
@@ -35,13 +35,13 @@ struct RootNavigator: Navigator {
     private func handleRootAction(_ action: Root.Action) {
         switch action {
         case .didReachFortyTwo:
-            applyStateConfiguration(.showPage)
+            applyTargetState(.showPage)
         }
     }
 }
 
 extension RootNavigator {
-    enum StateConfiguration: Hashable {
+    enum TargetState: Hashable {
         case reset
         case showPage
     }