+++ Sources/Puddles/Queryables/QueryableWithInput.swift
@@ -6,7 +6,7 @@ import SwiftUI
 /// An example use case would be a boolean coming from a confirmation dialog view. First, create a property of the desired data type:
 ///
 /// ```swift
-/// @QueryableItem<String, Bool> var deletionConfirmation
+/// @QueryableWithInput<String, Bool> var deletionConfirmation
 /// ```
 ///
 /// Then, use one of the `queryable` prefixed presentation modifiers to show the deletion confirmation. Here, we use an alert:
@@ -27,7 +27,7 @@ import SwiftUI
 ///     }
 /// ```
 ///
-/// To actually present the alert and await the boolean result, call ``Puddles/QueryableItem/Trigger/query(with:)`` on the ``Puddles/QueryableItem`` property.
+/// To actually present the alert and await the boolean result, call ``Puddles/QueryableWithInput/Trigger/query(with:)`` on the ``Puddles/QueryableWithInput`` property.
 /// This will activate the alert presentation which can then resolve the query in its completion handler.
 ///
 /// ```swift
@@ -37,19 +37,19 @@ import SwiftUI
 /// } catch {}
 /// ```
 ///
-/// When the Task that calls ``Puddles/QueryableItem/Trigger/query(with:)`` is cancelled, the suspended query will also cancel and deactivate (i.e. close) the wrapped navigation presentation.
+/// When the Task that calls ``Puddles/QueryableWithInput/Trigger/query(with:)`` is cancelled, the suspended query will also cancel and deactivate (i.e. close) the wrapped navigation presentation.
 /// In that case, a ``Puddles/QueryCancellationError`` error is thrown.
 ///
 /// For more information, see <doc:05-Queryable>.
 @propertyWrapper
-public struct QueryableItem<Item, Result>: DynamicProperty where Item: Sendable, Result: Sendable {
+public struct QueryableWithInput<Item, Result>: DynamicProperty where Item: Sendable, Result: Sendable {
 
     /// A representation of the `Queryable` property wrapper type.
     public struct Trigger {
 
-        /// A binding to the `item` state inside the `@QueryableItem` property wrapper.
+        /// A binding to the `item` state inside the `@QueryableWithInput` property wrapper.
         ///
-        /// This is used internally inside ``Puddles/QueryableItem/Wrapper/query()``.
+        /// This is used internally inside ``Puddles/QueryableWithInput/Wrapper/query()``.
         var item: Binding<Item?>
 
         /// A pointer to the ``Puddles/QueryResolver`` object that is used to resolve the query.
@@ -64,7 +64,7 @@ public struct QueryableItem<Item, Result>: DynamicProperty where Item: Sendable,
 
         /// A pointer to the `Buffer` object type.
         ///
-        /// This is used internally inside ``Puddles/QueryableItem/Wrapper/query()``.
+        /// This is used internally inside ``Puddles/QueryableWithInput/Wrapper/query()``.
         private var buffer: QueryBuffer<Result>
 
         /// A representation of the `Queryable` property wrapper type. This can be passed to `queryable` prefixed presentation modifiers, like `queryableSheet`.
@@ -82,7 +82,7 @@ public struct QueryableItem<Item, Result>: DynamicProperty where Item: Sendable,
         ///
         /// This method will suspend for as long as the query is unanswered and not cancelled. When the parent Task is cancelled, this method will immediately cancel the query and throw a ``Puddles/QueryCancellationError`` error.
         ///
-        /// Creating multiple queries at the same time will cause a query conflict which is resolved using the ``Puddles/QueryConflictPolicy`` defined in the initializer of ``Puddles/QueryableItem``. The default policy is ``Puddles/QueryConflictPolicy/cancelPreviousQuery``.
+        /// Creating multiple queries at the same time will cause a query conflict which is resolved using the ``Puddles/QueryConflictPolicy`` defined in the initializer of ``Puddles/QueryableWithInput``. The default policy is ``Puddles/QueryConflictPolicy/cancelPreviousQuery``.
         /// - Returns: The result of the query.
         public func query(with item: Item) async throws -> Result {
             return try await withTaskCancellationHandler {
@@ -123,10 +123,10 @@ public struct QueryableItem<Item, Result>: DynamicProperty where Item: Sendable,
         .init(item: $item, resolver: resolver, buffer: buffer)
     }
 
-    /// Internal helper type that stores and continues a `CheckedContinuation` created by calling ``Puddles/QueryableItem/Trigger/query()``.
+    /// Internal helper type that stores and continues a `CheckedContinuation` created by calling ``Puddles/QueryableWithInput/Trigger/query()``.
     private var buffer: QueryBuffer<Result>
 
-    /// Helper type to hide implementation details of ``Puddles/QueryableItem``.
+    /// Helper type to hide implementation details of ``Puddles/QueryableWithInput``.
     /// This type exposes convenient methods to answer (i.e. complete) a query.
     private var resolver: QueryResolver<Result> {
         .init(